BoundaryResponse and function to mdmsUtil
